+++ examples/fastapi-mcp-server/.env.example
@@ -1,0 +1,2 @@
+DESCOPE_PROJECT_ID=<your Descope project id>
+DESCOPE_API_BASE_URL=https://api.descope.com
