+++ examples/fastapi-mcp-server/app/auth/auth_config.py
@@ -10,8 +10,7 @@ class Settings(BaseSettings):
     
     @property
     def issuer_candidates(self) -> str:
-        # the 'iss' field is the Descope Project ID.
-        return [self.descope_project_id, f'https://api.descope.com/v1/apps/{self.descope_project_id}']
+        return [f'{self.descope_api_base_url}/v1/apps/{self.descope_project_id}', self.descope_project_id, ]
     
     @property
     def audience(self) -> str:
